@@ -20,23 +20,23 @@
   },
   "devDependencies": {
     "@skypack/package-check": "^0.2.2",
-    "@types/jest": "^27.4.1",
-    "@typescript-eslint/eslint-plugin": "^5.14.0",
-    "@typescript-eslint/parser": "^5.14.0",
+    "@types/jest": "^29.2.3",
+    "@typescript-eslint/eslint-plugin": "^5.44.0",
+    "@typescript-eslint/parser": "^5.44.0",
     "cross-env": "^7.0.3",
-    "es-check": "^6.2.1",
-    "eslint": "^8.10.0",
+    "es-check": "^7.0.1",
+    "eslint": "^8.28.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-config-react-app": "^7.0.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.25.4",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "jest": "v28.0.0-alpha.4",
-    "lerna": "^4.0.0",
-    "prettier": "^2.5.1",
-    "ts-jest": "latest",
-    "typescript": "^4.6.2"
+    "eslint-plugin-promise": "^6.1.1",
+    "jest": "^29.3.1",
+    "lerna": "^6.0.3",
+    "prettier": "^2.7.1",
+    "ts-jest": "^29.0.3",
+    "typescript": "^4.9.3"
   },
   "version": "0.0.0",
   "packageManager": "yarn@3.2.0"
