@@ -27,11 +27,11 @@
     "./dist"
   ],
   "dependencies": {
-    "@aws-sdk/client-cognito-identity": "^3.54.0",
-    "@aws-sdk/client-s3": "^3.54.0",
-    "@aws-sdk/credential-provider-cognito-identity": "^3.54.0"
+    "@aws-sdk/client-cognito-identity": "^3.215.0",
+    "@aws-sdk/client-s3": "^3.215.0",
+    "@aws-sdk/credential-provider-cognito-identity": "^3.215.0"
   },
   "devDependencies": {
-    "typescript": "^4.6.2"
+    "typescript": "^4.9.3"
   }
 }
