@@ -14,29 +14,29 @@
     "README.md"
   ],
   "dependencies": {
-    "@aws-sdk/client-cloudformation": "^3.54.0",
-    "@aws-sdk/client-s3": "^3.54.0",
+    "@aws-sdk/client-cloudformation": "^3.215.0",
+    "@aws-sdk/client-s3": "^3.215.0",
     "chalk": "^4.1.2",
     "cross-spawn": "^7.0.3",
-    "dotenv": "^16.0.0",
-    "fs-extra": "^10.0.1",
-    "glob": "^7.2.0",
-    "glob-promise": "^4.2.2",
+    "dotenv": "^16.0.3",
+    "fs-extra": "^10.1.0",
+    "glob": "^8.0.3",
+    "glob-promise": "^5.0.0",
     "json2csv": "^5.0.7",
     "lodash-es": "^4.17.21",
-    "mime-types": "^2.1.34",
+    "mime-types": "^2.1.35",
     "s3-sync-client": "^1.5.1",
-    "yargs": "^17.3.1"
+    "yargs": "^17.6.2"
   },
   "devDependencies": {
     "@types/cross-spawn": "^6.0.2",
     "@types/fs-extra": "^9.0.13",
     "@types/json2csv": "^5.0.3",
     "@types/mime-types": "^2.1.1",
-    "@types/node": "^17.0.21",
-    "@types/yargs": "^17.0.9",
-    "ts-node": "^10.7.0",
-    "typescript": "^4.6.2"
+    "@types/node": "^18.11.9",
+    "@types/yargs": "^17.0.13",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.3"
   },
   "scripts": {
     "start": "tsc",
