@@ -38,18 +38,18 @@
     "generate-schema": "ts-json-schema-generator --tsconfig tsconfig.json -p src/types.ts -t ConfigurationNoGenerics > src/schemas.json"
   },
   "dependencies": {
-    "core-js": "^3.21.1",
+    "core-js": "^3.26.1",
     "lodash-es": "^4.17.21",
-    "regenerator-runtime": "^0.13.9"
+    "regenerator-runtime": "^0.13.11"
   },
   "devDependencies": {
     "@types/deep-freeze": "^0.1.2",
-    "@types/jest": "^27.4.1",
+    "@types/jest": "^29.2.3",
     "@types/lodash-es": "^4.17.6",
-    "@types/node": "^17.0.21",
+    "@types/node": "^18.11.9",
     "deep-freeze": "^0.0.1",
-    "ts-json-schema-generator": "^0.98.0",
-    "typescript": "^4.6.2"
+    "ts-json-schema-generator": "^1.1.2",
+    "typescript": "^4.9.3"
   },
   "gitHead": "8d0f2af41121f70996e8b5803d006832b147fd86"
 }
